Updated collision, keeping one list of all blocks hit, changed deltaTime to float for slow-motion

+++ src/net/fybertech/ld29/GUI.java
@@ -71,7 +71,7 @@ public class GUI
 		childGUIs.add(child);
 	}
 	
-	public void update(int deltaTime)
+	public void update(float deltaTime)
 	{
 		while (Mouse.next())
 		{
